(1357) Feedback: create a folder for matomoCtlr
